// apps/portal/src/api/generated/model/questionOutcomeDto.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * HOA Manager API
 * The HOA Manager API description
 * OpenAPI spec version: 1.0
 */
import type { QuestionOutcomeDtoOutcome } from "./questionOutcomeDtoOutcome";

export interface QuestionOutcomeDto {
    questionId: string;
    title: string;
    outcome: QuestionOutcomeDtoOutcome;
    /** @nullable */
    winningOptionLabel: string | null;
}