// apps/portal/src/api/generated/model/questionOutcomeDtoOutcome.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * HOA Manager API
 * The HOA Manager API description
 * OpenAPI spec version: 1.0
 */

export type QuestionOutcomeDtoOutcome =
    (typeof QuestionOutcomeDtoOutcome)[keyof typeof QuestionOutcomeDtoOutcome];

export const QuestionOutcomeDtoOutcome = {
    APPROVED: "APPROVED",
    REJECTED: "REJECTED",
    WINNER: "WINNER",
    NOT_DECIDED: "NOT_DECIDED",
} as const;